Move `NoReplyMixin` to main `__init__.py` file
